api/requirements.txt
fastapi==0.115.7
openai_whisper==20240930
